requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1

# Database
supabase==2.10.0
asyncpg==0.30.0
sqlalchemy==2.0.35

# AI/ML
google-generativeai==0.8.3
pillow==11.0.0

# Async
aiohttp==3.10.10
aiofiles==24.1.0

# File Monitoring
watchdog==5.0.3

# Utilities
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
httpx==0.27.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Logging
loguru==0.7.2

# Task Queue (optional)
celery==5.4.0
redis==5.2.0